requirements.txt
i2cdevice
drv8830
paho-mqtt
